+++ bentotui.go
@@ -5,8 +5,8 @@ import (
 	"github.com/cloudboy-jh/bentotui/core"
 	"github.com/cloudboy-jh/bentotui/core/router"
 	"github.com/cloudboy-jh/bentotui/core/theme"
-	"github.com/cloudboy-jh/bentotui/ui/components/footer"
-	"github.com/cloudboy-jh/bentotui/ui/components/header"
+	"github.com/cloudboy-jh/bentotui/ui/containers/footer"
+	"github.com/cloudboy-jh/bentotui/ui/containers/header"
 )
 
 func New(opts ...app.Option) *app.Model {